Testando abstract repository
