+++ app/Repositories/CustomerRepository.php
@@ -5,7 +5,7 @@ namespace App\Repositories;
 use App\Models\Customer;
 use Illuminate\Database\Eloquent\Builder;
 
-class CustomerRepository {
+class CustomerRepository extends AbstractRepository {
 
     protected $model = Customer::class;
 
@@ -58,17 +58,5 @@ class CustomerRepository {
         $response = $this->find($id);
         $response->delete();
         return $response;
-    }
-
-    public function paginate(Builder $filter, $params)
-    {
-        $page = $params['page'] ?? 1;
-        $per_page = $params['per_page'] ?? 200;
-
-        if (!$filter) {
-            $filter = $this->model;
-        }
-
-        return $filter->paginate($per_page, ['*'], 'page', $page);
     }
 }
